Stream HCP loader and default Experiment data to float32
Two fixes from the weekend mancova_vba OOM diagnosis (kernel global OOM
at 16:40 on May 8, 30 workers × ~6.5 GB RSS on a 188 GB box):

1. load_image_nii: two-pass stream-and-mask instead of dict-of-images
   (peak per worker drops from ~5.6 GB to one image + output y).

2. Float32 default for Experiment data, with optional dtype kwarg on
   the public factories (from_search / from_paths / from_gauss).
   Audit pass eliminates silent float64 promotion at:
   - mancova.decompose (np.eye → x.dtype)
   - get_mancova y.mean accumulator
   - Experiment.__init__ add_bias path
   - Experiment.permute / get_freed_lane
   - ExperimentScaled mean_orig / pre_scale
   - bootstrap_img noise

Net effect on the GLOW analysis hot loop: compute_llr_batched is 1.98×
faster at HCP100/27k vox (72 ms → 36 ms), y.nbytes halved exactly.
Production mancova precision (size ≥ 4, 100 subjects) holds: max
relative error 4e-5, NaN counts unchanged.

Three test tolerance updates accompany the dtype change (float32
working precision, not behavioral): ExperimentScaled cov-orthogonality
check, get_mancova vs reference primitives, and a forced-float64
sentinel for the singular-matrix LinAlgError guardrail test.

+++ glow/experiment/permute.py
@@ -33,8 +33,11 @@ def get_freed_lane(x, contrast, perm_idx):
     rng = np.random.default_rng(perm_idx)
     perm = np.argsort(rng.permutation(num_img))
 
-    # freed_lane = P @ (I - Q0) + Q0, where P @ M = M[perm_inv, :]
-    return (np.eye(num_img) - q0)[perm, :] + q0
+    # freed_lane = P @ (I - Q0) + Q0, where P @ M = M[perm_inv, :].
+    # Match q0's dtype on np.eye so the subtraction doesn't promote a
+    # float32 q0 to float64 (which would then propagate into the
+    # permute einsum against y).
+    return (np.eye(num_img, dtype=q0.dtype) - q0)[perm, :] + q0
 
 
 class NotEnoughPermutations(Warning):
